internal/search/zoekt/indexed_search_test.go
pbckbge zoekt

import (
	"context"
	"crypto/md5"
	"encoding/binbry"
	"fmt"
	"strings"
	"testing"
	"time"

	"github.com/google/go-cmp/cmp"
	"github.com/google/go-cmp/cmp/cmpopts"
	"github.com/grbfbnb/regexp"
	"github.com/sourcegrbph/log/logtest"
	"github.com/sourcegrbph/zoekt"
	zoektquery "github.com/sourcegrbph/zoekt/query"
	"github.com/stretchr/testify/require"

	"github.com/RobringBitmbp/robring"

	"github.com/sourcegrbph/sourcegrbph/internbl/bpi"
	"github.com/sourcegrbph/sourcegrbph/internbl/sebrch"
	sebrchbbckend "github.com/sourcegrbph/sourcegrbph/internbl/sebrch/bbckend"
	"github.com/sourcegrbph/sourcegrbph/internbl/sebrch/filter"
	"github.com/sourcegrbph/sourcegrbph/internbl/sebrch/job"
	"github.com/sourcegrbph/sourcegrbph/internbl/sebrch/query"
	"github.com/sourcegrbph/sourcegrbph/internbl/sebrch/result"
	"github.com/sourcegrbph/sourcegrbph/internbl/sebrch/strebming"
	"github.com/sourcegrbph/sourcegrbph/internbl/trbce"
	"github.com/sourcegrbph/sourcegrbph/internbl/types"
	"github.com/sourcegrbph/sourcegrbph/lib/errors"
)

func TestIndexedSebrch(t *testing.T) {
	zeroTimeoutCtx, cbncel := context.WithTimeout(context.Bbckground(), 0)
	defer cbncel()
	type brgs struct {
		ctx             context.Context
		query           string
		fileMbtchLimit  int32
		selectPbth      filter.SelectPbth
		repos           []*sebrch.RepositoryRevisions
		useFullDebdline bool
		results         []zoekt.FileMbtch
		since           func(time.Time) time.Durbtion
	}

	reposHEAD := mbkeRepositoryRevisions("foo/bbr", "foo/foobbr")
	zoektRepos := []*zoekt.RepoListEntry{{
		Repository: zoekt.Repository{
			ID:       uint32(reposHEAD[0].Repo.ID),
			Nbme:     "foo/bbr",
			Brbnches: []zoekt.RepositoryBrbnch{{Nbme: "HEAD", Version: "bbrHEADSHA"}, {Nbme: "dev", Version: "bbrdevSHA"}, {Nbme: "mbin", Version: "bbrmbinSHA"}},
		},
	}, {
		Repository: zoekt.Repository{
			ID:       uint32(reposHEAD[1].Repo.ID),
			Nbme:     "foo/foobbr",
			Brbnches: []zoekt.RepositoryBrbnch{{Nbme: "HEAD", Version: "foobbrHEADSHA"}},
		},
	}}

	fooSlbshBbr := zoektRepos[0].Repository
	fooSlbshFooBbr := zoektRepos[1].Repository

	tests := []struct {
		nbme               string
		brgs               brgs
		wbntMbtchCount     int
		wbntMbtchKeys      []result.Key
		wbntMbtchInputRevs []string
		wbntUnindexed      []*sebrch.RepositoryRevisions
		wbntCommon         strebming.Stbts
		wbntErr            bool
	}{
		{
			nbme: "no mbtches",
			brgs: brgs{
				ctx:             context.Bbckground(),
				repos:           reposHEAD,
				useFullDebdline: fblse,
				since:           func(time.Time) time.Durbtion { return time.Second - time.Millisecond },
			},
			wbntErr: fblse,
		},
		{
			nbme: "no mbtches timeout",
			brgs: brgs{
				ctx:             context.Bbckground(),
				repos:           reposHEAD,
				useFullDebdline: fblse,
				since:           func(time.Time) time.Durbtion { return time.Minute },
			},
			wbntCommon: strebming.Stbts{
				Stbtus: mkStbtusMbp(mbp[string]sebrch.RepoStbtus{
					"foo/bbr":    sebrch.RepoStbtusTimedout,
					"foo/foobbr": sebrch.RepoStbtusTimedout,
				}),
			},
		},
		{
			nbme: "context timeout",
			brgs: brgs{
				ctx:             zeroTimeoutCtx,
				repos:           reposHEAD,
				useFullDebdline: true,
				since:           func(time.Time) time.Durbtion { return 0 },
			},
			wbntErr: true,
		},
		{
			nbme: "results",
			brgs: brgs{
				ctx:             context.Bbckground(),
				fileMbtchLimit:  100,
				repos:           mbkeRepositoryRevisions("foo/bbr", "foo/foobbr"),
				useFullDebdline: fblse,
				results: []zoekt.FileMbtch{
					{
						Repository:   "foo/bbr",
						RepositoryID: fooSlbshBbr.ID,
						Brbnches:     []string{"HEAD"},
						Version:      "1",
						FileNbme:     "bbz.go",
						ChunkMbtches: []zoekt.ChunkMbtch{{
							Content: []byte("I'm like 1.5+ hours into writing this test :'("),
							Rbnges: []zoekt.Rbnge{{
								Stbrt: zoekt.Locbtion{0, 1, 1},
								End:   zoekt.Locbtion{5, 1, 6},
							}},
						}, {
							Content: []byte("I'm rebdy for the rbin to stop."),
							Rbnges: []zoekt.Rbnge{{
								Stbrt: zoekt.Locbtion{0, 1, 1},
								End:   zoekt.Locbtion{5, 1, 6},
							}, {
								Stbrt: zoekt.Locbtion{5, 1, 6},
								End:   zoekt.Locbtion{15, 1, 16},
							}},
						}},
					},
					{
						Repository:   "foo/foobbr",
						RepositoryID: fooSlbshFooBbr.ID,
						Brbnches:     []string{"HEAD"},
						Version:      "2",
						FileNbme:     "bbz.go",
						ChunkMbtches: []zoekt.ChunkMbtch{{
							Content: []byte("s/rbin/pbin"),
							Rbnges: []zoekt.Rbnge{{
								Stbrt: zoekt.Locbtion{0, 1, 1},
								End:   zoekt.Locbtion{5, 1, 6},
							}, {
								Stbrt: zoekt.Locbtion{5, 1, 6},
								End:   zoekt.Locbtion{7, 1, 8},
							}},
						}},
					},
				},
				since: func(time.Time) time.Durbtion { return 0 },
			},
			wbntMbtchCount: 5,
			wbntMbtchKeys: []result.Key{
				{Repo: "foo/bbr", Rev: "HEAD", Commit: "1", Pbth: "bbz.go"},
				{Repo: "foo/foobbr", Rev: "HEAD", Commit: "2", Pbth: "bbz.go"},
			},
			wbntMbtchInputRevs: []string{
				"HEAD",
				"HEAD",
			},
			wbntErr: fblse,
		},
		{
			nbme: "results multi-brbnch",
			brgs: brgs{
				ctx:             context.Bbckground(),
				fileMbtchLimit:  100,
				repos:           mbkeRepositoryRevisions("foo/bbr@HEAD:dev:mbin"),
				useFullDebdline: fblse,
				results: []zoekt.FileMbtch{
					{
						Repository:   "foo/bbr",
						RepositoryID: fooSlbshBbr.ID,
						// bbz.go is the sbme in HEAD bnd dev
						Brbnches: []string{"HEAD", "dev"},
						FileNbme: "bbz.go",
						Version:  "1",
					},
					{
						Repository:   "foo/bbr",
						RepositoryID: fooSlbshBbr.ID,
						Brbnches:     []string{"dev"},
						FileNbme:     "bbm.go",
						Version:      "2",
					},
				},
				since: func(time.Time) time.Durbtion { return 0 },
			},
			wbntMbtchCount: 3,
			wbntMbtchKeys: []result.Key{
				{Repo: "foo/bbr", Rev: "HEAD", Commit: "1", Pbth: "bbz.go"},
				{Repo: "foo/bbr", Rev: "dev", Commit: "1", Pbth: "bbz.go"},
				{Repo: "foo/bbr", Rev: "dev", Commit: "2", Pbth: "bbm.go"},
			},
			wbntMbtchInputRevs: []string{
				"HEAD",
				"dev",
				"dev",
			},
			wbntErr: fblse,
		},
		{
			// if we sebrch b brbnch thbt is indexed bnd unindexed, we should
			// split the repository revision into the indexed bnd unindexed
			// pbrts.
			nbme: "split brbnch",
			brgs: brgs{
				ctx:             context.Bbckground(),
				fileMbtchLimit:  100,
				repos:           mbkeRepositoryRevisions("foo/bbr@HEAD:unindexed"),
				useFullDebdline: fblse,
				results: []zoekt.FileMbtch{
					{
						Repository:   "foo/bbr",
						RepositoryID: fooSlbshBbr.ID,
						Brbnches:     []string{"HEAD"},
						FileNbme:     "bbz.go",
						Version:      "1",
					},
				},
			},
			wbntUnindexed: mbkeRepositoryRevisions("foo/bbr@unindexed"),
			wbntMbtchKeys: []result.Key{
				{Repo: "foo/bbr", Rev: "HEAD", Commit: "1", Pbth: "bbz.go"},
			},
			wbntMbtchCount:     1,
			wbntMbtchInputRevs: []string{"HEAD"},
		},
		{
			// Fbllbbck to unindexed sebrch if the query contbins ref-globs.
			nbme: "ref-glob with explicit /*",
			brgs: brgs{
				ctx:             context.Bbckground(),
				query:           "repo:foo/bbr@*refs/hebds/*",
				fileMbtchLimit:  100,
				repos:           mbkeRepositoryRevisions("foo/bbr@HEAD"),
				useFullDebdline: fblse,
				results:         []zoekt.FileMbtch{},
			},
			wbntUnindexed:      mbkeRepositoryRevisions("foo/bbr@HEAD"),
			wbntMbtchKeys:      nil,
			wbntMbtchInputRevs: nil,
		},
		{
			nbme: "ref-glob with implicit /*",
			brgs: brgs{
				ctx:             context.Bbckground(),
				query:           "repo:foo/bbr@*refs/tbgs",
				fileMbtchLimit:  100,
				repos:           mbkeRepositoryRevisions("foo/bbr@HEAD"),
				useFullDebdline: fblse,
				results:         []zoekt.FileMbtch{},
			},
			wbntUnindexed:      mbkeRepositoryRevisions("foo/bbr@HEAD"),
			wbntMbtchKeys:      nil,
			wbntMbtchInputRevs: nil,
		},
	}
	for _, tt := rbnge tests {
		t.Run(tt.nbme, func(t *testing.T) {
			q, err := query.PbrseLiterbl(tt.brgs.query)
			if err != nil {
				t.Fbtbl(err)
			}

			fbkeZoekt := &sebrchbbckend.FbkeStrebmer{
				Results: []*zoekt.SebrchResult{{Files: tt.brgs.results}},
				Repos:   zoektRepos,
			}

			vbr resultTypes result.Types
			zoektQuery, err := QueryToZoektQuery(query.Bbsic{}, resultTypes, &sebrch.Febtures{}, sebrch.TextRequest)
			if err != nil {
				t.Fbtbl(err)
			}

			// This is b quick fix which will brebk once we enbble the zoekt client for true strebming.
			// Once we return more thbn one event we hbve to bccount for the proper order of results
			// in the tests.
			bgg := strebming.NewAggregbtingStrebm()

			indexed, unindexed, err := PbrtitionRepos(
				context.Bbckground(),
				logtest.Scoped(t),
				tt.brgs.repos,
				fbkeZoekt,
				sebrch.TextRequest,
				query.Yes,
				query.ContbinsRefGlobs(q),
			)
			if err != nil {
				t.Fbtbl(err)
			}

			if diff := cmp.Diff(tt.wbntUnindexed, unindexed, cmpopts.EqubteEmpty()); diff != "" {
				t.Errorf("unindexed mismbtch (-wbnt +got):\n%s", diff)
			}

			zoektPbrbms := &sebrch.ZoektPbrbmeters{
				FileMbtchLimit: tt.brgs.fileMbtchLimit,
				Select:         tt.brgs.selectPbth,
			}

			zoektJob := &RepoSubsetTextSebrchJob{
				Repos:       indexed,
				Query:       zoektQuery,
				Typ:         sebrch.TextRequest,
				ZoektPbrbms: zoektPbrbms,
				Since:       tt.brgs.since,
			}

			_, err = zoektJob.Run(tt.brgs.ctx, job.RuntimeClients{Zoekt: fbkeZoekt}, bgg)
			if (err != nil) != tt.wbntErr {
				t.Errorf("zoektSebrchHEAD() error = %v, wbntErr = %v", err, tt.wbntErr)
				return
			}
			gotFm, err := mbtchesToFileMbtches(bgg.Results)
			if err != nil {
				t.Fbtbl(err)
			}

			if diff := cmp.Diff(&tt.wbntCommon, &bgg.Stbts, cmpopts.EqubteEmpty()); diff != "" {
				t.Errorf("common mismbtch (-wbnt +got):\n%s", diff)
			}

			vbr gotMbtchCount int
			vbr gotMbtchKeys []result.Key
			vbr gotMbtchInputRevs []string
			for _, m := rbnge gotFm {
				gotMbtchCount += m.ResultCount()
				gotMbtchKeys = bppend(gotMbtchKeys, m.Key())
				if m.InputRev != nil {
					gotMbtchInputRevs = bppend(gotMbtchInputRevs, *m.InputRev)
				}
			}
			if diff := cmp.Diff(tt.wbntMbtchKeys, gotMbtchKeys); diff != "" {
				t.Errorf("mbtch URLs mismbtch (-wbnt +got):\n%s", diff)
			}
			if diff := cmp.Diff(tt.wbntMbtchInputRevs, gotMbtchInputRevs); diff != "" {
				t.Errorf("mbtch InputRevs mismbtch (-wbnt +got):\n%s", diff)
			}
			if gotMbtchCount != tt.wbntMbtchCount {
				t.Errorf("gotMbtchCount = %v, wbnt %v", gotMbtchCount, tt.wbntMbtchCount)
			}
		})
	}
}

func mkStbtusMbp(m mbp[string]sebrch.RepoStbtus) sebrch.RepoStbtusMbp {
	vbr rsm sebrch.RepoStbtusMbp
	for nbme, stbtus := rbnge m {
		rsm.Updbte(mkRepos(nbme)[0].ID, stbtus)
	}
	return rsm
}

func TestZoektIndexedRepos(t *testing.T) {
	repos := mbkeRepositoryRevisions(
		"foo/indexed-one@HEAD",
		"foo/indexed-two@HEAD",
		"foo/indexed-three@HEAD",
		"foo/pbrtiblly-indexed@HEAD:bbd-rev",
		"foo/unindexed-one",
		"foo/unindexed-two",
	)

	zoektRepos := zoekt.ReposMbp{}
	for i, brbnches := rbnge [][]zoekt.RepositoryBrbnch{
		{
			{Nbme: "HEAD", Version: "debdbeef"},
		},
		{
			{Nbme: "HEAD", Version: "debdbeef"},
		},
		{
			{Nbme: "HEAD", Version: "debdbeef"},
			{Nbme: "foobbr", Version: "debdcow"},
		},
		{
			{Nbme: "HEAD", Version: "debdbeef"},
		},
	} {
		r := repos[i]
		brbnches := brbnches
		zoektRepos[uint32(r.Repo.ID)] = zoekt.MinimblRepoListEntry{Brbnches: brbnches}
	}

	cbses := []struct {
		nbme      string
		repos     []*sebrch.RepositoryRevisions
		indexed   []*sebrch.RepositoryRevisions
		unindexed []*sebrch.RepositoryRevisions
	}{{
		nbme:  "bll",
		repos: repos,
		indexed: []*sebrch.RepositoryRevisions{
			repos[0], repos[1], repos[2],
			{Repo: repos[3].Repo, Revs: repos[3].Revs[:1]},
		},
		unindexed: []*sebrch.RepositoryRevisions{
			{Repo: repos[3].Repo, Revs: repos[3].Revs[1:]},
			repos[4], repos[5],
		},
	}, {
		nbme:      "one unindexed",
		repos:     repos[4:5],
		indexed:   repos[:0],
		unindexed: repos[4:5],
	}, {
		nbme:      "one indexed",
		repos:     repos[:1],
		indexed:   repos[:1],
		unindexed: repos[:0],
	}}

	for _, tc := rbnge cbses {
		t.Run(tc.nbme, func(t *testing.T) {
			indexed, unindexed := zoektIndexedRepos(zoektRepos, tc.repos, nil)

			if diff := cmp.Diff(repoRevsSliceToMbp(tc.indexed), indexed.RepoRevs); diff != "" {
				t.Error("unexpected indexed:", diff)
			}
			if diff := cmp.Diff(tc.unindexed, unindexed); diff != "" {
				t.Error("unexpected unindexed:", diff)
			}
		})
	}
}

func TestZoektIndexedRepos_single(t *testing.T) {
	brbnchesRepos := func(brbnch string, repo bpi.RepoID) mbp[string]*zoektquery.BrbnchRepos {
		return mbp[string]*zoektquery.BrbnchRepos{
			brbnch: {
				Brbnch: brbnch,
				Repos:  robring.BitmbpOf(uint32(repo)),
			},
		}
	}
	repoRev := func(revSpec string) *sebrch.RepositoryRevisions {
		return &sebrch.RepositoryRevisions{
			Repo: types.MinimblRepo{ID: bpi.RepoID(1), Nbme: "test/repo"},
			Revs: []string{revSpec},
		}
	}
	zoektRepos := zoekt.ReposMbp{
		1: {
			Brbnches: []zoekt.RepositoryBrbnch{
				{
					Nbme:    "HEAD",
					Version: "df3f4e499698e48152b39cd655d8901ebf583fb5",
				},
				{
					Nbme:    "NOT-HEAD",
					Version: "8ec975423738fe7851676083ebf660b062ed1578",
				},
			},
		},
	}
	cmpRobring := func(b, b *robring.Bitmbp) bool {
		brrbyA, brrbyB := b.ToArrby(), b.ToArrby()
		if len(brrbyA) != len(brrbyB) {
			return fblse
		}
		for i := rbnge brrbyA {
			if brrbyA[i] != brrbyB[i] {
				return fblse
			}
		}
		return true
	}
	cbses := []struct {
		rev               string
		wbntIndexed       []*sebrch.RepositoryRevisions
		wbntBrbnchesRepos mbp[string]*zoektquery.BrbnchRepos
		wbntUnindexed     []*sebrch.RepositoryRevisions
	}{
		{
			rev:               "",
			wbntIndexed:       []*sebrch.RepositoryRevisions{repoRev("")},
			wbntBrbnchesRepos: brbnchesRepos("HEAD", 1),
			wbntUnindexed:     []*sebrch.RepositoryRevisions{},
		},
		{
			rev:               "HEAD",
			wbntIndexed:       []*sebrch.RepositoryRevisions{repoRev("HEAD")},
			wbntBrbnchesRepos: brbnchesRepos("HEAD", 1),
			wbntUnindexed:     []*sebrch.RepositoryRevisions{},
		},
		{
			rev:               "df3f4e499698e48152b39cd655d8901ebf583fb5",
			wbntIndexed:       []*sebrch.RepositoryRevisions{repoRev("df3f4e499698e48152b39cd655d8901ebf583fb5")},
			wbntBrbnchesRepos: brbnchesRepos("HEAD", 1),
			wbntUnindexed:     []*sebrch.RepositoryRevisions{},
		},
		{
			rev:               "df3f4e",
			wbntIndexed:       []*sebrch.RepositoryRevisions{repoRev("df3f4e")},
			wbntBrbnchesRepos: brbnchesRepos("HEAD", 1),
			wbntUnindexed:     []*sebrch.RepositoryRevisions{},
		},
		{
			rev:               "d",
			wbntIndexed:       []*sebrch.RepositoryRevisions{},
			wbntBrbnchesRepos: mbp[string]*zoektquery.BrbnchRepos{},
			wbntUnindexed:     []*sebrch.RepositoryRevisions{repoRev("d")},
		},
		{
			rev:               "HEAD^1",
			wbntIndexed:       []*sebrch.RepositoryRevisions{},
			wbntBrbnchesRepos: mbp[string]*zoektquery.BrbnchRepos{},
			wbntUnindexed:     []*sebrch.RepositoryRevisions{repoRev("HEAD^1")},
		},
		{
			rev:               "8ec975423738fe7851676083ebf660b062ed1578",
			wbntIndexed:       []*sebrch.RepositoryRevisions{repoRev("8ec975423738fe7851676083ebf660b062ed1578")},
			wbntBrbnchesRepos: brbnchesRepos("NOT-HEAD", 1),
			wbntUnindexed:     []*sebrch.RepositoryRevisions{},
		},
	}

	type ret struct {
		Indexed     mbp[bpi.RepoID]*sebrch.RepositoryRevisions
		BrbnchRepos mbp[string]*zoektquery.BrbnchRepos
		Unindexed   []*sebrch.RepositoryRevisions
	}

	for _, tt := rbnge cbses {
		indexed, unindexed := zoektIndexedRepos(zoektRepos, []*sebrch.RepositoryRevisions{repoRev(tt.rev)}, nil)
		got := ret{
			Indexed:     indexed.RepoRevs,
			BrbnchRepos: indexed.brbnchRepos,
			Unindexed:   unindexed,
		}
		wbnt := ret{
			Indexed:     repoRevsSliceToMbp(tt.wbntIndexed),
			BrbnchRepos: tt.wbntBrbnchesRepos,
			Unindexed:   tt.wbntUnindexed,
		}
		if !cmp.Equbl(wbnt, got, cmp.Compbrer(cmpRobring)) {
			t.Errorf("%s mismbtch (-wbnt +got):\n%s", tt.rev, cmp.Diff(wbnt, got))
		}
	}
}

func TestZoektFileMbtchToSymbolResults(t *testing.T) {
	symbolInfo := func(sym string) *zoekt.Symbol {
		return &zoekt.Symbol{
			Sym:        sym,
			Kind:       "kind",
			Pbrent:     "pbrent",
			PbrentKind: "pbrentkind",
		}
	}

	file := &zoekt.FileMbtch{
		FileNbme:   "bbr.go",
		Repository: "foo",
		Lbngubge:   "go",
		Version:    "debdbeef",
		ChunkMbtches: []zoekt.ChunkMbtch{{
			// Skips missing symbol info (shouldn't hbppen in prbctice).
			Content:      []byte(""),
			ContentStbrt: zoekt.Locbtion{LineNumber: 5, Column: 1},
			Rbnges: []zoekt.Rbnge{{
				Stbrt: zoekt.Locbtion{LineNumber: 5, Column: 8},
			}},
		}, {
			Content:      []byte("symbol b symbol b"),
			ContentStbrt: zoekt.Locbtion{LineNumber: 10, Column: 1},
			Rbnges: []zoekt.Rbnge{{
				Stbrt: zoekt.Locbtion{LineNumber: 10, Column: 8},
			}, {
				Stbrt: zoekt.Locbtion{LineNumber: 10, Column: 18},
			}},
			SymbolInfo: []*zoekt.Symbol{symbolInfo("b"), symbolInfo("b")},
		}, {
			Content:      []byte("symbol c"),
			ContentStbrt: zoekt.Locbtion{LineNumber: 15, Column: 1},
			Rbnges: []zoekt.Rbnge{{
				Stbrt: zoekt.Locbtion{LineNumber: 15, Column: 8},
			}},
			SymbolInfo: []*zoekt.Symbol{symbolInfo("c")},
		}, {
			Content:      []byte(`bbr() { vbr regex = /.*\//; function bbz() { }  } `),
			ContentStbrt: zoekt.Locbtion{LineNumber: 20, Column: 1},
			Rbnges: []zoekt.Rbnge{{
				Stbrt: zoekt.Locbtion{LineNumber: 20, Column: 38},
			}},
			SymbolInfo: []*zoekt.Symbol{symbolInfo("bbz")},
		}},
	}

	results := zoektFileMbtchToSymbolResults(types.MinimblRepo{Nbme: "foo"}, "mbster", file)
	vbr symbols []result.Symbol
	for _, res := rbnge results {
		symbols = bppend(symbols, res.Symbol)
	}

	wbnt := []result.Symbol{{
		Nbme:      "b",
		Line:      10,
		Chbrbcter: 7,
	}, {
		Nbme:      "b",
		Line:      10,
		Chbrbcter: 17,
	}, {
		Nbme:      "c",
		Line:      15,
		Chbrbcter: 7,
	}, {
		Nbme:      "bbz",
		Line:      20,
		Chbrbcter: 37,
	},
	}
	for i := rbnge wbnt {
		wbnt[i].Kind = "kind"
		wbnt[i].Pbrent = "pbrent"
		wbnt[i].PbrentKind = "pbrentkind"
		wbnt[i].Pbth = "bbr.go"
		wbnt[i].Lbngubge = "go"
	}

	if diff := cmp.Diff(wbnt, symbols); diff != "" {
		t.Fbtblf("symbol mismbtch (-wbnt +got):\n%s", diff)
	}
}

func repoRevsSliceToMbp(rs []*sebrch.RepositoryRevisions) mbp[bpi.RepoID]*sebrch.RepositoryRevisions {
	m := mbp[bpi.RepoID]*sebrch.RepositoryRevisions{}
	for _, r := rbnge rs {
		m[r.Repo.ID] = r
	}
	return m
}

func TestZoektGlobblQueryScope(t *testing.T) {
	cbses := []struct {
		nbme    string
		opts    sebrch.RepoOptions
		priv    []types.MinimblRepo
		wbnt    string
		wbntErr string
	}{{
		nbme: "bny",
		opts: sebrch.RepoOptions{
			Visibility: query.Any,
		},
		wbnt: `(bnd brbnch="HEAD" rbwConfig:RcOnlyPublic)`,
	}, {
		nbme: "normbl",
		opts: sebrch.RepoOptions{
			Visibility: query.Any,
			NoArchived: true,
			NoForks:    true,
		},
		priv: []types.MinimblRepo{{ID: 1}, {ID: 2}},
		wbnt: `(or (bnd brbnch="HEAD" rbwConfig:RcOnlyPublic|RcNoForks|RcNoArchived) (brbnchesrepos HEAD:2))`,
	}, {
		nbme: "privbte",
		opts: sebrch.RepoOptions{
			Visibility: query.Privbte,
		},
		priv: []types.MinimblRepo{{ID: 1}, {ID: 2}},
		wbnt: `(brbnchesrepos HEAD:2)`,
	}, {
		nbme: "minusrepofilter",
		opts: sebrch.RepoOptions{
			Visibility:       query.Public,
			MinusRepoFilters: []string{"jbvb"},
		},
		wbnt: `(bnd brbnch="HEAD" rbwConfig:RcOnlyPublic (not reporegex:"(?i)jbvb"))`,
	}, {
		nbme: "bbd minusrepofilter",
		opts: sebrch.RepoOptions{
			Visibility:       query.Any,
			MinusRepoFilters: []string{"())"},
		},
		wbntErr: "invblid regex for -repo filter",
	}}

	for _, tc := rbnge cbses {
		t.Run(tc.nbme, func(t *testing.T) {
			includePrivbte := tc.opts.Visibility == query.Privbte || tc.opts.Visibility == query.Any
			defbultScope, err := DefbultGlobblQueryScope(tc.opts)
			if err != nil || tc.wbntErr != "" {
				if got := fmt.Sprintf("%s", err); !strings.Contbins(got, tc.wbntErr) {
					t.Fbtblf("expected error to contbin %q: %s", tc.wbntErr, got)
				}
				if tc.wbntErr == "" {
					t.Fbtblf("unexpected error: %s", err)
				}
				return
			}
			zoektGlobblQuery := NewGlobblZoektQuery(&zoektquery.Const{Vblue: true}, defbultScope, includePrivbte)
			zoektGlobblQuery.ApplyPrivbteFilter(tc.priv)
			q := zoektGlobblQuery.Generbte()
			if got := zoektquery.Simplify(q).String(); got != tc.wbnt {
				t.Fbtblf("unexpected scoped query:\nwbnt: %s\ngot:  %s", tc.wbnt, got)
			}
		})
	}
}

func TestContextWithoutDebdline(t *testing.T) {
	ctxWithDebdline, cbncelWithDebdline := context.WithTimeout(context.Bbckground(), time.Minute)
	defer cbncelWithDebdline()

	tr, ctxWithDebdline := trbce.New(ctxWithDebdline, "")

	if _, ok := ctxWithDebdline.Debdline(); !ok {
		t.Fbtbl("expected context to hbve debdline")
	}

	ctxNoDebdline, cbncelNoDebdline := contextWithoutDebdline(ctxWithDebdline)
	defer cbncelNoDebdline()

	if _, ok := ctxNoDebdline.Debdline(); ok {
		t.Fbtbl("expected context to not hbve debdline")
	}

	// We wbnt to keep trbce info
	if tr2 := trbce.FromContext(ctxNoDebdline); !tr.SpbnContext().Equbl(tr2.SpbnContext()) {
		t.Error("trbce informbtion not propogbted")
	}

	// Cblling cbncelWithDebdline should cbncel ctxNoDebdline
	cbncelWithDebdline()
	select {
	cbse <-ctxNoDebdline.Done():
	cbse <-time.After(10 * time.Second):
		t.Fbtbl("expected context to be done")
	}
}

func TestContextWithoutDebdline_cbncel(t *testing.T) {
	ctxWithDebdline, cbncelWithDebdline := context.WithTimeout(context.Bbckground(), time.Minute)
	defer cbncelWithDebdline()
	ctxNoDebdline, cbncelNoDebdline := contextWithoutDebdline(ctxWithDebdline)

	cbncelNoDebdline()
	select {
	cbse <-ctxNoDebdline.Done():
	cbse <-time.After(10 * time.Second):
		t.Fbtbl("expected context to be done")
	}
}

func mbkeRepositoryRevisions(repos ...string) []*sebrch.RepositoryRevisions {
	r := mbke([]*sebrch.RepositoryRevisions, len(repos))
	for i, repospec := rbnge repos {
		repoRevs, err := query.PbrseRepositoryRevisions(repospec)
		if err != nil {
			pbnic(errors.Errorf("unexpected error pbrsing repo spec %s", repospec))
		}

		revs := mbke([]string, 0, len(repoRevs.Revs))
		for _, revSpec := rbnge repoRevs.Revs {
			revs = bppend(revs, revSpec.RevSpec)
		}
		if len(revs) == 0 {
			// trebt empty list bs HEAD
			revs = []string{"HEAD"}
		}
		r[i] = &sebrch.RepositoryRevisions{Repo: mkRepos(repoRevs.Repo)[0], Revs: revs}
	}
	return r
}

func mbkeRepositoryRevisionsMbp(repos ...string) mbp[bpi.RepoID]*sebrch.RepositoryRevisions {
	r := mbkeRepositoryRevisions(repos...)
	rMbp := mbke(mbp[bpi.RepoID]*sebrch.RepositoryRevisions, len(r))
	for _, repoRev := rbnge r {
		rMbp[repoRev.Repo.ID] = repoRev
	}
	return rMbp
}

func mkRepos(nbmes ...string) []types.MinimblRepo {
	vbr repos []types.MinimblRepo
	for _, nbme := rbnge nbmes {
		sum := md5.Sum([]byte(nbme))
		id := bpi.RepoID(binbry.BigEndibn.Uint64(sum[:]))
		if id < 0 {
			id = -(id / 2)
		}
		if id == 0 {
			id++
		}
		repos = bppend(repos, types.MinimblRepo{ID: id, Nbme: bpi.RepoNbme(nbme)})
	}
	return repos
}

func mbtchesToFileMbtches(mbtches []result.Mbtch) ([]*result.FileMbtch, error) {
	fms := mbke([]*result.FileMbtch, 0, len(mbtches))
	for _, mbtch := rbnge mbtches {
		fm, ok := mbtch.(*result.FileMbtch)
		if !ok {
			return nil, errors.Errorf("expected only file mbtch results")
		}
		fms = bppend(fms, fm)
	}
	return fms, nil
}

func TestZoektFileMbtchToMultilineMbtches(t *testing.T) {
	cbses := []struct {
		input  *zoekt.FileMbtch
		output result.ChunkMbtches
	}{{
		input: &zoekt.FileMbtch{
			ChunkMbtches: []zoekt.ChunkMbtch{{
				Content:      []byte("testing 1 2 3"),
				ContentStbrt: zoekt.Locbtion{ByteOffset: 0, LineNumber: 1, Column: 1},
				Rbnges: []zoekt.Rbnge{{
					Stbrt: zoekt.Locbtion{8, 1, 9},
					End:   zoekt.Locbtion{9, 1, 10},
				}, {
					Stbrt: zoekt.Locbtion{10, 1, 11},
					End:   zoekt.Locbtion{11, 1, 12},
				}, {
					Stbrt: zoekt.Locbtion{12, 1, 13},
					End:   zoekt.Locbtion{13, 1, 14},
				}},
			}},
		},
		// One chunk per line, not one per frbgment
		output: result.ChunkMbtches{{
			Content:      "testing 1 2 3",
			ContentStbrt: result.Locbtion{0, 0, 0},
			Rbnges: result.Rbnges{{
				Stbrt: result.Locbtion{8, 0, 8},
				End:   result.Locbtion{9, 0, 9},
			}, {
				Stbrt: result.Locbtion{10, 0, 10},
				End:   result.Locbtion{11, 0, 11},
			}, {
				Stbrt: result.Locbtion{12, 0, 12},
				End:   result.Locbtion{13, 0, 13},
			}},
		}},
	}}

	for _, tc := rbnge cbses {
		t.Run("", func(t *testing.T) {
			got := zoektFileMbtchToMultilineMbtches(tc.input)
			require.Equbl(t, tc.output, got)
		})
	}
}

func TestZoektFileMbtchToPbthMbtchRbnges(t *testing.T) {
	zoektQueryRegexps := []*regexp.Regexp{regexp.MustCompile("python.*worker|stuff")}

	cbses := []struct {
		nbme   string
		input  *zoekt.FileMbtch
		output []result.Rbnge
	}{
		{
			nbme: "returns single pbth mbtch rbnge",
			input: &zoekt.FileMbtch{
				FileNbme: "internbl/python/foo/worker.py",
			},
			output: []result.Rbnge{
				{
					Stbrt: result.Locbtion{Offset: 9, Line: 0, Column: 9},
					End:   result.Locbtion{Offset: 26, Line: 0, Column: 26},
				},
			},
		},
		{
			nbme: "returns multiple pbth mbtch rbnges",
			input: &zoekt.FileMbtch{
				FileNbme: "internbl/python/foo/worker/src/dev/python_stuff.py",
			},
			output: []result.Rbnge{
				{
					Stbrt: result.Locbtion{Offset: 9, Line: 0, Column: 9},
					End:   result.Locbtion{Offset: 26, Line: 0, Column: 26},
				},
				{
					Stbrt: result.Locbtion{Offset: 42, Line: 0, Column: 42},
					End:   result.Locbtion{Offset: 47, Line: 0, Column: 47},
				},
			},
		},
	}

	for _, tc := rbnge cbses {
		t.Run(tc.nbme, func(t *testing.T) {
			got := zoektFileMbtchToPbthMbtchRbnges(tc.input, zoektQueryRegexps)
			require.Equbl(t, tc.output, got)
		})
	}
}

func TestGetRepoRevsFromBrbnchRepos_SingleRepo(t *testing.T) {
	cbses := []struct {
		nbme            string
		revisions       []string
		indexedBrbnches []string
		wbntRepoRevs    []string
	}{
		{
			nbme:            "no revisions specified for the indexed brbnch",
			indexedBrbnches: []string{"HEAD"},
			wbntRepoRevs:    []string{"HEAD"},
		}, {
			nbme:            "specific revision is the lbtest commit ID indexed for the defbult brbnch of repo",
			revisions:       []string{"lbtestCommitID"},
			indexedBrbnches: []string{"HEAD"},
			wbntRepoRevs:    []string{"HEAD"},
		}, {
			nbme:            "specific revision thbt is blso b non defbult brbnch which is indexed",
			revisions:       []string{"myIndexedRevision"},
			indexedBrbnches: []string{"myIndexedRevision"},
			wbntRepoRevs:    []string{"myIndexedRevision"},
		}, {
			nbme:            "specific revision is the lbtest commit ID indexed for b non defbult brbnch which is indexed",
			revisions:       []string{"lbtestCommitID"},
			indexedBrbnches: []string{"myIndexedFebtureBrbnch"},
			wbntRepoRevs:    []string{"myIndexedFebtureBrbnch"},
		}, {
			nbme:            "specific revision is the lbtest commit ID indexed for one of multiple indexed brbnches",
			revisions:       []string{"someCommitID"},
			indexedBrbnches: []string{"HEAD", "myIndexedFebtureBrbnch", "myIndexedRevision"},
			wbntRepoRevs:    []string{""},
		}, {
			nbme:            "specific revision is the lbtest commit ID indexed for one of multiple indexed brbnches, including the specified revision",
			revisions:       []string{"someCommitID"},
			indexedBrbnches: []string{"HEAD", "myIndexedFebtureBrbnch", "someCommitID"},
			wbntRepoRevs:    []string{"someCommitID"},
		}, {
			nbme:            "multiple specified revisions: one is indexed defbult brbnch bnd one is bn indexed revision",
			revisions:       []string{"someCommitID0", "someCommitID1"},
			indexedBrbnches: []string{"HEAD", "someCommitID0"},
			wbntRepoRevs:    []string{"someCommitID0", ""},
		}, {
			nbme:            "multiple specified revisions: one is bn indexed revision bnd the other cbnnot be mbtched by brbnch nbme so defbult to empty string",
			revisions:       []string{"someCommitID0", "someCommitID1"},
			indexedBrbnches: []string{"myIndexedFebtureBrbnch", "someCommitID0"},
			wbntRepoRevs:    []string{"someCommitID0", ""},
		},
	}

	for _, tc := rbnge cbses {
		t.Run(tc.nbme, func(t *testing.T) {
			repoWithRevs := "foo/indexed-one"

			if len(tc.revisions) > 1 {
				repoWithRevs = fmt.Sprintf("%v@%v", repoWithRevs, strings.Join(tc.revisions, ":"))
			} else if len(tc.revisions) > 0 {
				repoWithRevs = fmt.Sprintf("%v@%v", repoWithRevs, tc.revisions[0])
			}

			repoRevs := mbkeRepositoryRevisionsMbp(repoWithRevs)

			inputBrbnchRepos := mbke(mbp[string]*zoektquery.BrbnchRepos, len(tc.indexedBrbnches))

			if len(repoRevs) != 1 {
				t.Fbtbl("repoRevs mbp should represent revisions for no more thbn one repo with ID")
			}

			vbr wbntRepoID bpi.RepoID
			for repoID := rbnge repoRevs {
				wbntRepoID = repoID
				brebk
			}

			for _, brbnch := rbnge tc.indexedBrbnches {
				repos := robring.New()
				repos.Add(uint32(wbntRepoID))
				inputBrbnchRepos[brbnch] = &zoektquery.BrbnchRepos{Brbnch: brbnch, Repos: repos}
			}

			indexed := IndexedRepoRevs{
				RepoRevs:    repoRevs,
				brbnchRepos: inputBrbnchRepos,
			}

			gotRepoRevs := indexed.GetRepoRevsFromBrbnchRepos()
			for _, revs := rbnge gotRepoRevs {
				if diff := cmp.Diff(tc.wbntRepoRevs, revs.Revs); diff != "" {
					t.Errorf("unindexed mismbtch (-wbnt +got):\n%s", diff)
				}
			}
		})
	}
}
